+++ openfisca_core/columns.py
@@ -34,11 +34,10 @@ class Column(object):
     definition_period = None
     # json_type = None  # Defined in sub-classes
     label = None
-    law_reference = None  # Either a single reference or a list of references
     name = None
     start = None
     survey_only = False
-    url = None
+    reference = None  # Either a single reference or a list of references
     val_type = None
 
     def __init__(
@@ -49,10 +48,9 @@ class Column(object):
             entity = None,
             function = None,
             label = None,
-            law_reference = None,
             start = None,
             survey_only = False,
-            url = None,
+            reference = None,
             val_type = None
             ):
         if cerfa_field is not None:
@@ -64,16 +62,14 @@ class Column(object):
             self.end = end
         if function is not None:
             self.function = function
-        if law_reference is not None:
-            self.law_reference = law_reference
         if label is not None:
             self.label = label
         if start is not None:
             self.start = start
         if survey_only:
             self.survey_only = True
-        if url is not None:
-            self.url = url
+        if reference is not None:
+            self.reference = reference
         if val_type is not None and val_type != self.val_type:
             self.val_type = val_type
         self.is_neutralized = False
@@ -173,8 +169,8 @@ class Column(object):
             self_json['start'] = start
         if self.survey_only:
             self_json['survey_only'] = self.survey_only
-        if self.url is not None:
-            self_json['url'] = self.url
+        if self.reference is not None:
+            self_json['reference'] = self.reference
         if self.val_type is not None:
             self_json['val_type'] = self.val_type
         return self_json