+++ openfisca_core/reforms.py
@@ -13,7 +13,7 @@ from .taxbenefitsystems import TaxBenefitSystem
 def compose_reforms(reforms, tax_benefit_system):
     """
     Compose reforms: the first reform is built with the given base tax-benefit system,
-    then each one is built with the previous one as the reference.
+    then each one is built with the previous one as the baseline.
     """
     def compose_reforms_reducer(memo, reform):
         reformed_tbs = reform(memo)
@@ -25,39 +25,39 @@ def compose_reforms(reforms, tax_benefit_system):
 class Reform(TaxBenefitSystem):
     name = None
 
-    def __init__(self, reference):
-        self.reference = reference
-        self._legislation_json = reference.get_legislation()
-        self.compact_legislation_by_instant_cache = reference.compact_legislation_by_instant_cache
-        self.column_by_name = reference.column_by_name.copy()
-        self.decomposition_file_path = reference.decomposition_file_path
-        self.Scenario = reference.Scenario
+    def __init__(self, baseline):
+        self.baseline = baseline
+        self._legislation_json = baseline.get_legislation()
+        self.compact_legislation_by_instant_cache = baseline.compact_legislation_by_instant_cache
+        self.column_by_name = baseline.column_by_name.copy()
+        self.decomposition_file_path = baseline.decomposition_file_path
+        self.Scenario = baseline.Scenario
         self.key = unicode(self.__class__.__name__)
         if not hasattr(self, 'apply'):
             raise Exception("Reform {} must define an `apply` function".format(self.key))
         self.apply()
 
     def __getattr__(self, attribute):
-        return getattr(self.reference, attribute)
+        return getattr(self.baseline, attribute)
 
     @property
     def full_key(self):
         key = self.key
         assert key is not None, 'key was not set for reform {} (name: {!r})'.format(self, self.name)
-        if self.reference is not None and hasattr(self.reference, 'key'):
-            reference_full_key = self.reference.full_key
-            key = u'.'.join([reference_full_key, key])
+        if self.baseline is not None and hasattr(self.baseline, 'key'):
+            baseline_full_key = self.baseline.full_key
+            key = u'.'.join([baseline_full_key, key])
         return key
 
     def modify_legislation_json(self, modifier_function):
         """
-        Copy the reference TaxBenefitSystem legislation_json attribute and return it.
+        Copy the baseline TaxBenefitSystem legislation_json attribute and return it.
         Used by reforms which need to modify the legislation_json, usually in the build_reform() function.
         Validates the new legislation.
         """
-        reference_legislation_json = self.reference.get_legislation()
-        reference_legislation_json_copy = copy.deepcopy(reference_legislation_json)
-        reform_legislation_json = modifier_function(reference_legislation_json_copy)
+        baseline_legislation_json = self.baseline.get_legislation()
+        baseline_legislation_json_copy = copy.deepcopy(baseline_legislation_json)
+        reform_legislation_json = modifier_function(baseline_legislation_json_copy)
         assert reform_legislation_json is not None, \
             'modifier_function {} in module {} must return the modified legislation_json'.format(
                 modifier_function.__name__,